Moved Game to model
I moved the Game class in the Game Library page to a new model file inside a new folder called Models.  This folder is meant to define all important classes, like chat messages, events, games, etc.
