+++ lib/pages/game_library_page.dart
@@ -1,9 +1,9 @@
 // ignore_for_file: prefer_const_constructors
 import 'dart:ffi';
-
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
+import '../models/game.dart';
 
 class GameLibrary extends StatefulWidget {
   const GameLibrary({super.key});
@@ -130,27 +130,3 @@ class _GameLibraryState extends State<GameLibrary> {
   }
 }
 
-class Game {
-  final int id;
-  final String name;
-  final String released;
-  final Array platforms;
-  final String background_image;
-
-  Game(
-      {required this.id,
-      required this.name,
-      required this.released,
-      required this.platforms,
-      required this.background_image});
-
-  factory Game.fromJson(Map<String, dynamic> json) {
-    return Game(
-      id: json['id'],
-      name: json['name'],
-      released: json['released'],
-      platforms: json['genres'],
-      background_image: json['background_image'],
-    );
-  }
-}
